requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiounittest==1.4.1
asn1crypto==1.5.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
bcrypt==3.2.0
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
dnspython==2.2.1
email-validator==1.2.0
Faker==13.4.0
Flask==2.1.1
Flask-Login==0.6.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.33.2
fpdf==1.7.2
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
multidict==6.0.2
oscrypto==1.3.0
Pillow==9.1.0
pycparser==2.21
pyHanko==0.12.1
pyhanko-certvalidator==0.19.5
pyOpenSSL==22.0.0
python-barcode==0.13.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.36
tzdata==2022.1
tzlocal==4.2
uharfbuzz==0.19.0
uritools==4.0.0
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.14.0
WTForms==3.0.1
yarl==1.7.2
zipp==3.8.0
